Update - Added new disassembly view to function pointer type.

+++ ReClass/CNodeByte.h
@@ -9,7 +9,7 @@ public:
 
 	virtual void Update( HotSpot& Spot );
 
-	virtual int GetMemorySize( ) { return sizeof( unsigned char ); }
+	virtual ULONG GetMemorySize( ) { return sizeof( unsigned char ); }
 
 	virtual int Draw( ViewInfo& View, int x, int y );
 };
