+++ ReClass/CNodeFloat.h
@@ -9,7 +9,7 @@ public:
 
 	virtual void Update( HotSpot& Spot );
 
-	virtual int GetMemorySize( ) { return sizeof( float ); }
+	virtual ULONG GetMemorySize( ) { return sizeof( float ); }
 
 	virtual int Draw( ViewInfo& View, int x, int y );
 };
