+++ ReClass/CNodeQWORD.h
@@ -9,7 +9,7 @@ public:
 
 	virtual void Update( HotSpot& Spot );
 
-	virtual int GetMemorySize( ) { return sizeof( unsigned long long ); }
+	virtual ULONG GetMemorySize( ) { return sizeof( unsigned long long ); }
 
 	virtual int Draw( ViewInfo& View, int x, int y );
 };
